Squats screen

+++ app/src/main/java/com/example/superfit/presentation/view/screens/exercises/crunch/CrunchScreenModel.kt
@@ -10,6 +10,6 @@ sealed class CrunchScreenIntent {
 
 @Immutable
 data class CrunchScreenState(
-    val repeatsCount: Float? = null,
+    val repeatsCount: Int? = null,
     val navigateToSuccessScreen: Boolean? = null
 )
